chore: move PeopleCard to cards
