+++ website-frontend/src/routes/people/+page.svelte
@@ -4,7 +4,7 @@
 	import CardPanel from '$lib/components/panel/CardPanel.svelte';
 	import FilterBar from '$lib/components/filter/FilterBar.svelte';
 	import LoadMore from '$lib/components/buttons/LoadMore.svelte';
-	import PeopleCard from '$lib/components/people/PeopleCard.svelte';
+	import PeopleCard from '$lib/components/cards/PeopleCard.svelte';
 	export let data;
 
 	$: ({ people, people_overview, position_filters, laboratory_filters } = data);
